Update tests to work with new Otel packages
