+++ examples/express/package.json
@@ -12,8 +12,8 @@
   "license": "ISC",
   "devDependencies": {
     "@autometrics/typescript-plugin": "^0.5.4",
-    "@opentelemetry/exporter-prometheus": "^0.43.0",
-    "@opentelemetry/sdk-metrics": "^1.17.0",
+    "@opentelemetry/exporter-prometheus": "^0.45.0",
+    "@opentelemetry/sdk-metrics": "^1.18.0",
     "@types/express": "^4.17.16",
     "@types/node": "^18.6.5",
     "typescript": "^5.2.2"
