+++ packages/autometrics/tests/buildInfo/customBuildInfo.test.ts
@@ -1,14 +1,14 @@
 import { assertMatch, assertStringIncludes } from "$std/assert/mod.ts";
 
-import { autometrics, recordBuildInfo } from "../mod.ts";
+import { autometrics, recordBuildInfo } from "../../mod.ts";
 import {
   BRANCH_LABEL,
   COMMIT_LABEL,
   REPOSITORY_URL_LABEL,
   SERVICE_NAME_LABEL,
   VERSION_LABEL,
-} from "../src/constants.ts";
-import { collectAndSerialize, stepWithMetricReader } from "./testUtils.ts";
+} from "../../src/constants.ts";
+import { collectAndSerialize, stepWithMetricReader } from "../testUtils.ts";
 
 Deno.test("Build info tests", async (t) => {
   await stepWithMetricReader(
