+++ packages/autometrics/tests/temporality/intervalDeltaTemporality.test.ts
@@ -1,0 +1,31 @@
+import { assert, assertEquals } from "$std/assert/mod.ts";
+
+import { MetricData } from "$otel/sdk-metrics";
+import { AggregationTemporalityPreference } from "../../src/exporter-otlp-http/mod.ts";
+import { testWithTemporality } from "../testUtils.ts";
+
+const assertMetricDataHasValue =
+  (value: number) => (data: MetricData | undefined) => {
+    assertEquals(data?.dataPoints.length, 1);
+    assertEquals(data?.dataPoints[0].value, value);
+  };
+
+const assertMetricDataIsEmpty = () => (data: MetricData | undefined) => {
+  if (data) {
+    assertEquals(data.dataPoints.length, 0);
+  } else {
+    assert(true, "no data is fine, we want it to be empty");
+  }
+};
+
+Deno.test("Temporality tests", async (t) => {
+  await t.step(
+    "clears metrics when they are pushed at an interval and delta temporality is used",
+    testWithTemporality({
+      temporalityPreference: AggregationTemporalityPreference.DELTA,
+      pushInterval: 50,
+      assertBeforePush: assertMetricDataHasValue(1),
+      assertAfterPush: assertMetricDataIsEmpty(),
+    }),
+  );
+});
